changes in tests according to review suggestions
